syntax modification

+++ in_tta/DecodeFile.h
@@ -72,7 +72,7 @@ public:
 
 	int				SetFileName(const char *filename);
 	const char	   *GetFileName(){return FileName.c_str();}
-	int				GetSamples(BYTE *buffer, long buffersize, int *current_bitrate);
+	int				GetSamples(BYTE *buffer, size_t buffersize, int *current_bitrate);
 
 	int				GetPaused(){return paused;}
 	void			SetPaused(int p){paused = p;}
